+++ contracts/interfaces/IMaidCoin.sol
@@ -6,24 +6,32 @@ import "./IRatio.sol";
 import "./IMasterCoin.sol";
 
 interface IMaidCoin is IERC20 {
-	
-    event Mint(address indexed to, uint amount);
-    event Burn(address indexed from, uint amount);
-    
+    event Mint(address indexed to, uint256 amount);
+    event Burn(address indexed from, uint256 amount);
+
     function ratio() external view returns (IRatio);
+
     function masterCoin() external view returns (IMasterCoin);
-    
-    function initialSupply() external view returns (uint);
-    function maidCorpAccReward() external view returns (uint);
-    function nurseRaidAccReward() external view returns (uint);
-    
+
+    function initialSupply() external view returns (uint256);
+
+    function maidCorpAccReward() external view returns (uint256);
+
+    function nurseRaidAccReward() external view returns (uint256);
+
     function maidCorp() external view returns (address);
+
     function cloneNurse() external view returns (address);
+
     function maid() external view returns (address);
+
     function nurseRaid() external view returns (address);
-    
-    function mintForMaidCorp() external returns (uint);
-    function mintForCloneNurse() external returns (uint);
+
+    function mintForMaidCorp() external returns (uint256);
+
+    function mintForCloneNurse() external returns (uint256);
+
     function mintForCloneNurseDestruction(address to, uint256 amount) external;
-    function burn(address from, uint amount) external;
+
+    function burn(address from, uint256 amount) external;
 }
