+++ contracts/interfaces/IRNG.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.5;
 
 interface IRNG {
-    function generateRandomNumber(uint seed) external returns (uint);
+    function generateRandomNumber(uint256 seed) external returns (uint256);
 }
